fix: :bug: Final bug fixes
Final bug fixes

+++ AutoTallerManager.Application/DTOs/RegistrarHorasRealesDto.cs
@@ -2,5 +2,5 @@ namespace AutoTallerManager.Application.DTOs;
 
 public class RegistrarHorasRealesDto
 {
-    public int HorasReales { get; set; }
+    public decimal HorasReales { get; set; }
 }
